Code cleanup

+++ api/test/setupTest.js
@@ -4,21 +4,31 @@ const mongoose = require("mongoose");
 require("dotenv").config();
 
 let dropDatabase = function(done) {
-  mongoose.connection.db
-    .dropDatabase()
-    .then(() => {
-      done();
-    })
-    .catch(err => {
-      console.error(err);
 
-      if (err.code === 26) {
-        console.log("Ignore error mongoose 26");
+  let dropAfterTest = false;
+
+  if(dropAfterTest)
+  {
+    mongoose.connection.db
+      .dropDatabase()
+      .then(() => {
         done();
-      } else {
-        throw err;
-      }
-    });
+      })
+      .catch(err => {
+        console.error(err);
+  
+        if (err.code === 26) {
+          console.log("Ignore error mongoose 26");
+          done();
+        } else {
+          throw err;
+        }
+      });
+  }
+  else{
+    //Don't drop the database
+    done();
+  }
 };
 
 //Establish connection to mongoDB
